+++ wmserver/src/window_controller.cpp
@@ -282,7 +282,7 @@ void WindowController::ProcessDisplayChange(DisplayId displayId, DisplayStateCha
             windowRoot_->NotifyDisplayChange(abstractDisplay);
 
             // TODO: Remove 'sysBarWinId_' after SystemUI resize 'systembar'
-            uint32_t width = abstractDisplay->GetWidth();
+            uint32_t width = static_cast<uint32_t>(abstractDisplay->GetWidth());
             uint32_t height = abstractDisplay->GetHeight() * SYSTEM_BAR_HEIGHT_RATIO;
             Rect newRect = { 0, 0, width, height };
             ResizeRect(sysBarWinId_[WindowType::WINDOW_TYPE_STATUS_BAR], newRect, WindowSizeChangeReason::DRAG);
